requirement.txt
mysql-connector-python
apscheduler
python-dotenv
logging
colorama
requests